Update to Dread Elves
Fixed and issue with Cult of Yema General.
It’s now possible to upgrade Acolytes and Medusae to chosen of Yema
with a Cult of Yema General.
Dancers of Yema are now possible to take with a Cult of Yema general.
Executioners Chosen of Nabh upgrade is now no longer free.
Updated rules text for Divine Blessings.