Numerous Fixes for VC and OK
Magic Items (Heroes)
- Mithril Mail should be 25 points, down from 50
- King Slayer seems to be missing for both lords and heroes

Lords
Necromancer Lord
- base cost should be 170, down from 175
Vampire Lord
- Stormcaller power should be 65pts, up from 50

Core
- Skeletons: Technically Shields come as standard, but thats OCD talking

Special
- Skirmisher Option for Vampiric spawn should really be a toggle option,
not a multiple choice.

ITEMS:
(Im checking this weapons on the lords lists)
Skullsplitter costs 50p and now is 40p.
Kingslayer weapon is missing from weapons list.
Axe of battle costs 30p and now is 20p.
Beastbane Halbert cost 30p and now is 25p.
Armor of fortune costs 30p and now is 25p.
Mithril Mail costs 30p, and now is 35 for Lords and 25 for heroes,
Flying carpet costs 50p, now is 40p on lords and 30p on heroes.
Wizard's hood costs 45p and now is 40p.
Book of thousand incantations is avaliable on lords and has to be
erased.
Essence of a free mind costs 100p and has to be 40p on lords, and 30p on
heroes.

(Now on heroes list)
Axe of battle costs 35p and has to be 25p.
Beast-bane Halberd costs 30p and has to be 25p.
Skullsplitter costs 50p and has to be 40p.
Kingslayer is missing from the heroes weapons list.
Armor of fortune costs 30p and has to be 25p.
Mithril mail costs 30p and has to be 25p.
Yeti furs costs 15p and has to be 10p.
Flying carpet costs 50p and has to be 30p.
Book of thousand incantations is avaliable and has to be erased.
Book of arcane power costs 50p and now is 35p.
Essence of a free mind costs 50p and now is 40p on lords and 30p on
heroes.
Tome of arcane lore costs 25p and now is 25 on lords and 15 on heroes.

STANDARDS:
Aether icon costs 40p and now is 30p.
Standard of alternate realities is avaliable and now is not. (also is on
the veteran magic standards list)

Lords:
Great Shaman:
Paths of Alchemy and Fire are not avaliable from the paths list.
The Greater Fire blessing now is not avaliable to choose from Special
rules.
On "view roster" his "WS" and "A" is 4 and now is 3.
Spinesplitter Big name is not avaliable on Big Names list.

Heroes:
Kahn:
CC weapons Iron fist cost 10p and now is 15p.

Shaman:
BIg name Spinesplitter is not avaliable on big names list.
Special rule Fire blessing is not avaliable on Special rules list.
CC weapons Paired weapons (it says aditional hand weapon) costs 2p and
is 3p.

Mammoth hunter:
On the "view roster" Hunting spear description apears as Brace of ogre
pistols.
CC weapons Iron fist costs 10p and is 15.
CC weapons Lance is not avaliable and now is and costs 15p.

Scraplings:
When you changes Throwing Weapons, it apears an error saying that
scraplings must have 10 more selections of throwing weapons (minimum 10)
(on the Roster Errors window).