Major Undying Dynasty Changes
Lord of the Barrow Legion:
Option for barding and lances added to skeleton cavalry when selected.
Option for Halberds for skeleton chariots, also hide light troops when
selected.

Commander of the Terracotta Army:
From what I could read and what was mentioned on the forum, "All other
NON-UNDEAD construct units and characters must be upgraded". All the
characters ARE undead so do not get +1T and -1 Initiative?
Might need a rules clarification on this.